Robotiq_gripper/with_OP/minial_whatsapp_bot/pyproject.toml
[project]
name = "my-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.17.0",
    "websockets>=16.0",
    "python-dotenv>=1.0.0",
    "opencv-python",
    "depthai",
    "pillow>=10.0.0",
]
